+++ main.go
@@ -1,10 +1,9 @@
 /*
 Copyright © 2025 NAME HERE <EMAIL ADDRESS>
-
 */
 package main
 
-import "BookServer/cmd"
+import "github.com/Sabnaj-42/BookServer-API/cmd"
 
 func main() {
 	cmd.Execute()
